+++ CHANGES.md
@@ -9,6 +9,11 @@
 * First stable release (TODO)
 
 **0.15.0**
+* Complete full `jau::fs::mount()` and `umount()`: 
+  - `mount()` for block devices etc (non loop-device image)
+  - `umount()` of mount-point
+  - MountFlags and UnmountFlags detailed for GNU/Linux
+  - Java binding and a manual unit test with TestsudoFileUtils02.
 * `ByteInStream_URL` available() and read() await `url_header_sync` for `m_content_size` and blocks at read if available
   - `ByteInStream_Feed::read()` blocks as well with known `m_content_size`, but has no `url_header_sync` mechanism
 * Introduce `url_header_sync` object, used for async `read_url_stream()` and hence `ByteInStream_URL`